generic actor and actor sheet
